+++ crates/wdl-lint/src/rules/snake_case.rs
@@ -140,6 +140,8 @@ impl Rule for SnakeCaseRule {
             SyntaxKind::StructDefinitionNode,
             SyntaxKind::TaskDefinitionNode,
             SyntaxKind::WorkflowDefinitionNode,
+            SyntaxKind::InputSectionNode,
+            SyntaxKind::OutputSectionNode,
             SyntaxKind::BoundDeclNode,
             SyntaxKind::UnboundDeclNode,
         ])